レビュー対応: グループファイルを個別ファイルに分割
- labor-outsourcing-section.tsx → labor-cost-section.tsx + outsourcing-cost-section.tsx
- development-equipment-section.tsx → development-cost-section.tsx + equipment-allocation-section.tsx
- logistics-electricity-section.tsx → logistics-cost-section.tsx + electricity-cost-section.tsx + fees-cost-section.tsx
- ファイル名とエクスポートの一致を確保
