+++ src/app/_components/cost/cost-tab.tsx
@@ -8,9 +8,13 @@ import { Button } from "@/components/ui/button"
 import type { AppData } from "@/lib/types"
 
 import { CostSummarySection } from "./sections/cost-summary-section"
-import { DevelopmentCostSection, EquipmentAllocationSection } from "./sections/development-equipment-section"
-import { LaborCostSection, OutsourcingCostSection } from "./sections/labor-outsourcing-section"
-import { LogisticsCostSection, ElectricityCostSection, FeesCostSection } from "./sections/logistics-electricity-section"
+import { DevelopmentCostSection } from "./sections/development-cost-section"
+import { ElectricityCostSection } from "./sections/electricity-cost-section"
+import { EquipmentAllocationSection } from "./sections/equipment-allocation-section"
+import { FeesCostSection } from "./sections/fees-cost-section"
+import { LaborCostSection } from "./sections/labor-cost-section"
+import { LogisticsCostSection } from "./sections/logistics-cost-section"
+import { OutsourcingCostSection } from "./sections/outsourcing-cost-section"
 import { PackagingCostSection } from "./sections/packaging-cost-section"
 import { ProfitSimulationSection } from "./sections/profit-simulation-section"
